requirements.txt
gym==0.19.0
ray==1.9.2
Jinja2
colorama==0.4.4
pddlpy==0.1.9
setproctitle==1.2.2
alabaster==0.7.12
glcontext==2.3.4
Sphinx==4.2.0
python-dateutil==2.8.2
pyparsing==2.4.7
numpy~=1.20.3
Babel==2.9.1
docutils==0.17.1
matplotlib
Pygments==2.10.0
protobuf==3.18.0
seaborn==0.11.2
procgen
torch-geometric
typed-ast<1.5,>=1.4.0
six~=1.15.0
tensorflow
glfw<2.0.0,>=1.8.6
MarkupSafe==2.0.1
Pillow==9.0.1
sklearn==0.0
plantcv==3.0.4
pandas>==1.1.0
scipy==1.5.0
tornado==4.5.1
squaternion==0.3.2
pyquaternion==0.9.9
networkx
open3d
opencv-contrib-python==4.5.3.56
cloudpickle==1.6.0
cffi==1.14.6
requests==2.21.0
Cython==0.29.24
flowvis==0.1.4
pyglet==1.5.21
redis==3.5.3
msgpack==1.0.2
lz4==3.1.3
GPUtil==1.4.0
dm-tree==0.1.6
ffmpeg==1.4
stable-baselines3==1.3.
clingo==5.5.0.post3
gym-minigrid==1.0.2
clorm==1.3.4
griddly==1.2.19
wandb==0.12.7
flow_vis==0.1
hvplot==0.7.3
holoviews==1.14.7
clingo
Clorm==1.3.4
torchtext==0.11.2
pygame==2.1.2